# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for the Master Class education platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List platform users
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/UserRole"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/UserStatus"
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a platform user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a platform user
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a platform user
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "204":
          description: User deleted
  /users/{userId}/password:
    patch:
      operationId: resetUserPassword
      tags: [users]
      summary: Set a new password for a platform user
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPasswordUpdate"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordUpdateResult"
        "404":
          description: User not found
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List classes
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ClassStatus"
      responses:
        "200":
          description: Classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /classes/{classId}:
    patch:
      operationId: updateClass
      tags: [classes]
      summary: Update a class
      parameters:
        - $ref: "#/components/parameters/ClassId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassUpdate"
      responses:
        "200":
          description: Updated class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete a class
      parameters:
        - $ref: "#/components/parameters/ClassId"
      responses:
        "204":
          description: Class deleted
  /classes/{classId}/assignments:
    get:
      operationId: getClassAssignments
      tags: [classes]
      summary: Get all assignments for a class
      parameters:
        - $ref: "#/components/parameters/ClassId"
      responses:
        "200":
          description: Class assignments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassAssignments"
  /classes/{classId}/teacher:
    put:
      operationId: assignClassTeacher
      tags: [classes]
      summary: Assign a teacher to a class
      parameters:
        - $ref: "#/components/parameters/ClassId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherAssignmentInput"
      responses:
        "200":
          description: Updated class assignments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassAssignments"
  /classes/{classId}/modules:
    put:
      operationId: assignClassModules
      tags: [classes]
      summary: Assign modules to a class
      parameters:
        - $ref: "#/components/parameters/ClassId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModulesAssignmentInput"
      responses:
        "200":
          description: Updated class assignments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassAssignments"
  /classes/{classId}/students:
    put:
      operationId: assignClassStudents
      tags: [classes]
      summary: Assign students to a class
      parameters:
        - $ref: "#/components/parameters/ClassId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentsAssignmentInput"
      responses:
        "200":
          description: Updated class assignments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassAssignments"
...
components:
  parameters:
    Search:
      name: search
      in: query
      schema:
        type: string
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
    ClassId:
      name: classId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    UserRole:
      type: string
      enum: [admin, teacher, student]
    UserStatus:
      type: string
      enum: [active, invited, suspended]
    ClassStatus:
      type: string
      enum: [active, draft, archived]
    User:
      type: object
      required: [id, name, email, role, status, classesCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/UserRole"
        status:
          $ref: "#/components/schemas/UserStatus"
        avatar:
          type: ["string", "null"]
        classesCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    UserInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          writeOnly: true
        role:
          $ref: "#/components/schemas/UserRole"
        status:
          $ref: "#/components/schemas/UserStatus"
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/UserRole"
        status:
          $ref: "#/components/schemas/UserStatus"
    UserPasswordUpdate:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 8
          writeOnly: true
        signOutOfOtherSessions:
          type: boolean
          default: true
    PasswordUpdateResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    Class:
      type: object
      required: [id, name, level, academicYear, status, studentCount, moduleCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: string
        academicYear:
          type: string
        status:
          $ref: "#/components/schemas/ClassStatus"
        teacher:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            name:
              type: string
        studentCount:
          type: integer
        moduleCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ClassInput:
      type: object
      required: [name, level, academicYear]
      properties:
        name:
          type: string
          minLength: 2
        level:
          type: string
        academicYear:
          type: string
        status:
          $ref: "#/components/schemas/ClassStatus"
    ClassUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        level:
          type: string
        academicYear:
          type: string
        status:
          $ref: "#/components/schemas/ClassStatus"
    Module:
      type: object
      required: [id, slug, name, shortDescription, color, activityCount, editorUrl, repositoryUrl]
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        shortDescription:
          type: string
        color:
          type: string
        activityCount:
          type: integer
        editorUrl:
          type: string
          format: uri
        repositoryUrl:
          type: string
          format: uri
...
    AssignedUser:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
    ClassAssignments:
      type: object
      required: [classId, teacher, students, modules]
      properties:
        classId:
          type: integer
        teacher:
          type: ["object", "null"]
          $ref: "#/components/schemas/AssignedUser"
        students:
          type: array
          items:
            $ref: "#/components/schemas/AssignedUser"
        modules:
          type: array
          items:
            $ref: "#/components/schemas/Module"
    TeacherAssignmentInput:
      type: object
      required: [teacherId]
      properties:
        teacherId:
          type: integer
    ModulesAssignmentInput:
      type: object
      required: [moduleIds]
      properties:
        moduleIds:
          type: array
          items:
            type: integer
    StudentsAssignmentInput:
      type: object
      required: [studentIds]
      properties:
        studentIds:
          type: array
          items:
            type: integer
...