# lib/api-zod/src/generated/types/classStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Master Class education platform
 * OpenAPI spec version: 0.1.0
 */

export type ClassStatus = typeof ClassStatus[keyof typeof ClassStatus];


export const ClassStatus = {
  active: 'active',
  draft: 'draft',
  archived: 'archived',
} as const;