# lib/api-zod/src/generated/types/module.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Master Class education platform
 * OpenAPI spec version: 0.1.0
 */

export interface Module {
  id: number;
  slug: string;
  name: string;
  shortDescription: string;
  color: string;
  activityCount: number;
  editorUrl: string;
  repositoryUrl: string;
}