# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Master Class education platform
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  avatar?: string | null;
  classesCount: number;
  createdAt: Date;
}