# lib/api-zod/src/generated/types/userPasswordUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Master Class education platform
 * OpenAPI spec version: 0.1.0
 */

export interface UserPasswordUpdate {
  /** @minLength 8 */
  password: string;
  signOutOfOtherSessions?: boolean;
}