# lib/api-zod/src/generated/types/userStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Master Class education platform
 * OpenAPI spec version: 0.1.0
 */

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  active: 'active',
  invited: 'invited',
  suspended: 'suspended',
} as const;